lims/ui: add CommonMultiPageDialog and downgrade react-scripts version
Summary:
Added CommonMultiPageDialog (an MUI version of MultiPageModal which was
bootstrap), and downgraded react-scripts version because it tried to
mess with z-index which was causing issues for our app.
Also updated dialog such that buttons can be on left or right.
Fix T9101
Fix T9105

Test Plan:
{F251895}
{F251894}
{F251914}

Reviewers: jsingh, jzhao, nsawas

Reviewed By: jsingh

Subscribers: jsingh

Maniphest Tasks: T9101, T9105

Differential Revision: https://phabricator.grailbio.com/D15705

fbshipit-source-id: 6f2a3d0