automation: change typeahead initial message
Summary:
Users have voiced confusion over the current automation typeahead,
as it is unclear that the component can search for existing instruments
asynchronously.

CommonTypeahead now allows for a custom message when using the
`async` select type. Without a custom message the component defaults
to an initial message, indicating for the user to begin typing
in order to see results.

Test Plan: {F348552}

Reviewers: jsingh, mbates

Reviewed By: jsingh

Subscribers: jsingh

Differential Revision: https://phabricator.grailbio.com/D17207

fbshipit-source-id: dc264d7