ui: create common tabbed-card
Summary:
New designs for a page with a flush card header with header tabs
have been created and require a new common component to support them.

`TabbedCard` creates a card header that allows for multiple tabs to be created
to switch between pages, and includes styling that are specific to the
designs for the component.

Fix T10159

Reviewers: jsingh, margi

Reviewed By: jsingh

Subscribers: jsingh

Maniphest Tasks: T10159

Differential Revision: https://phabricator.grailbio.com/D17286

fbshipit-source-id: 67d2726