ui: set size for checkbox column
Summary:
PagedTable and SimpleTable columns will vary in width depending on the content.
The checkbox column however should remain a standard width as the size of
the checkboxes do not change. This revision sets the checkbox width to be a set
size.

Fix T9996

Test Plan: {F355085}

Reviewers: jsingh, aahn

Reviewed By: jsingh

Maniphest Tasks: T9996

Differential Revision: https://phabricator.grailbio.com/D17304

fbshipit-source-id: 404a93d