ui/components: spread otherprops to common components
Summary:
The common UI library should be able to handle custom props passed in by consumers,
and should not be limited to the predefined props that are currently used.

This revision implements spread operators to apply any other props that may be defined
in the component where they are necessary.

Fix T10009

Reviewers: jsingh, jzhao

Reviewed By: jsingh, jzhao

Maniphest Tasks: T10009

Differential Revision: https://phabricator.grailbio.com/D17453

fbshipit-source-id: fb60389