ui/components: allow specification of custom date formats
Summary: This change allows the specification of a custom date format for the DateInput and DateTimeInput components. Specifying a date format affects the way that a date selected in the picker appears in the field and also changes the behavior of typed dates.

Test Plan: Specify a custom date format and pick a date from the picker. Ensure that the selected date is displayed according to the format specified.

Reviewers: jzhao, jsingh, O13 UI Common, nsawas

Reviewed By: jsingh, O13 UI Common, nsawas

Subscribers: mward

Differential Revision: https://phabricator.grailbio.com/D19073

fbshipit-source-id: 7a19ec1