lims/ui: add reagent lot detail page replacing modal
Summary:
The modal for reagent lot information did not fit on the screen. Per
the design document, this adds the first rendition of a reagent lot
detail page. After conversation with Minyong, we are scrapping the
"Batches" tab, so this diff currently adds tabs to a page even though
there is only one tab (the "Details") tab, but this is acceptable as it
will provide flexibility for including future information.

Screenshots:
{F479525}

{F479524}

{F479523}

{F479522}

Fix T11434
Fix T11389
Fix T11356

Test Plan: Cypress tests coming in next revision.

Reviewers: jsingh, margi, skong, O12 LIMS UI, O13 UI Common, nsawas

Reviewed By: jsingh, O12 LIMS UI, O13 UI Common, nsawas

Subscribers: minyongis, mz, nsawas, jrosenfield, alice

Maniphest Tasks: T11356, T11434, T11389

Differential Revision: https://phabricator.grailbio.com/D19178

fbshipit-source-id: d40e088