ui/lims: remove incomplete page links
Summary:
It is unclear if these pages will be complete by the
cut time for 1.20.0. This diff is to remove these links
if these pages do not get implemented in time. Placeholder
links should not be shown if a user cannot interact with it.

Will be watching for changes to T10071 and T10185

Edit: In order to fix unit tests, revert change from D19838 in
package.json files

Test Plan:
Ran locally and active links are unaffected. Inactive links are
hidden appropriately.

Reviewers: rhuang, jsingh, jrosenfield, jzhao, O13 UI Common, nsawas

Reviewed By: jsingh, O13 UI Common, nsawas

Differential Revision: https://phabricator.grailbio.com/D19970

fbshipit-source-id: 1da2e7b