lims: make study sample reference ID editable
Summary:
Study sample reference ID should be editable if the sample has not been
used in batch processing. The edit functionality is moved from a modal
that appears on the study sample table page to an inline editor on
the study sample detail page.

Cypress tests coming next.

This is the "OverrideWarningCreatingStudySamplesModal":
{F572912}

And this is the "OverrideWarningEditingStudySamplesModal":
{F572907}

Fix T11790

Test Plan: `go test`

Reviewers: alice, nsawas, rhuang, O10 LIMS Server, O12 LIMS UI, O13 UI Common, jsingh

Reviewed By: alice, nsawas, rhuang, O10 LIMS Server, O12 LIMS UI, O13 UI Common, jsingh

Subscribers: jsingh

Maniphest Tasks: T11790

Differential Revision: https://phabricator.grailbio.com/D19944

fbshipit-source-id: 266fcb2