lims/ui: change default instrument count
Summary:
This revision adds larger counts to pagination for `PagedTable` and changes
the default count for the `Instruments` table to use it.

This also fixes the compilation warning that `UiSpec` and `UiSpecRow` were defined
after already being used in the `wait-ui-spec.jsx` file.

Test Plan:
Check UI for pagination numbers
Check console for no compilation warnings

Reviewers: O14 Skynet UI, O13 UI Common, O12 LIMS UI, jsingh, lkong, mbates

Reviewed By: O14 Skynet UI, O13 UI Common, O12 LIMS UI, jsingh, mbates

Subscribers: aahn, jsingh, lkong

Differential Revision: https://phabricator.grailbio.com/D21994

fbshipit-source-id: cdd9c9b