instruments: fix instruments details page styling
Summary:
The instruments details page grows in length as the length of the timeline
graph grows. When the timeline is empty the cards are also not the same height.

This revision fixes this issue by updating the TimelieCard as necessary, and implementing
the proper styling fixes on the instruments datshboard.

Fix T15532

Test Plan:
Test on styleguidist
Test on instruments UI
Check Timelines in LIMS to make sure they were not affected improperly.

{F726517}

Reviewers: O13 UI Common, O12 LIMS UI, jsingh, ecarrel, lkong, margi

Reviewed By: O13 UI Common, O12 LIMS UI, ecarrel

Subscribers: jzhao, jsingh

Maniphest Tasks: T15532

Differential Revision: https://phabricator.grailbio.com/D24575

fbshipit-source-id: 56ae70a