lims/ui: make LargeTableColumn header have dynamic height
Summary:
`react-virtualized` tables incorporate the header into the table, and
fix the height of the header, perhaps to ensure that the table is a
fixed defined height, including the header. Secondarily, depending on
number of rows, the scroll bar width may affect placement of header
items, so it has to optionally add padding.
We don't require the table to be a certain height. Move header rendering
to be outside of the table.
Fix T16322

Test Plan:
Styleguidist
Test in batch panels
{F743969}

{F743968}

Reviewers: O13 UI Common, jzhao, nsawas

Reviewed By: O13 UI Common, jzhao

Subscribers: nsawas

Maniphest Tasks: T16322

Differential Revision: https://phabricator.grailbio.com/D25516

fbshipit-source-id: 93a9047