lims: rename lot number to lot unit number when relevant
Summary:
This revision replaces all relevant uses of "lot number" with
"lot unit number", so as to clearly differentiate which code refers to
a specific unit of a reagent lot and which code refers to just a
reagent lot but not a specific unit. This will make it easier to avoid
bugs like the one fixed in D24202, because a field purported to
reference a reagent lot number but really referenced a reagent lot unit
number.

What makes this awkward (and these bugs hard to spot) is that LIMS
still has to support old, unitless reagent lots. Consequently,
"lot unit number", throughout the codebase and UI, will now refer
either to:
- A reagent lot unit, e.g. "L00001-01" in the case that the "L00001"
  reagent lot was created after LIMS 1.24.0 when units were released OR
- A reagent lot, e.g. "L00001" in the case that the "L00001" reagent
  lot was created before LIMS 1.24.0 when units were released.

In general, this can be thought of as a 1:many relationship, a subset
of which are just a 1:1 relationship. Open to other suggestions to make
this less confusing, but we can't just give units to all lots because we
still need to support all the existing reagents in the lab that
already have labels printed on them without suffixes. (We could add
units to them and have LIMS automatically append "-01" to them
whenever they are scanned in, but in my opinion that is more confusing.
Additionally, we'd have to rewrite history by appending "-01" to all
existing lot numbers in the `BatchOperationReagent` table)

See https://docs.google.com/document/d/1riSxo_LutoFE47sUnwG_14ifbj1el7DMDfuOdwSd2f8/edit?disco=AAAACbubwNw&ts=5c184390&usp_dm=false
for an overview of the differentiation between reagent lot units and
reagent lots.

I did this by searching for both "lot number" and "lotnumber"
(case-insensitively) and refactoring all mentions which should instead
reference the lot unit number.

Ref T14279
Fix T15996

Test Plan:
- `go test`
- Tested various behavior manually in the UI: committing a batch,
  searching for reagent lots, searching for batch reagents, viewing
  table of reagent lots, editing a reagent lot, creating a reagent
  lot, printing a reagent lot, printing a reagent lot with units.

Reviewers: nsawas, rhuang, stimokhin, alice, O11 Skynet, O10 LIMS Server, O13 UI Common, O12 LIMS UI, jsingh, mbates, lkong

Reviewed By: nsawas, O11 Skynet, O10 LIMS Server, O13 UI Common, O12 LIMS UI, jsingh, lkong

Subscribers: lkong, jsingh

Maniphest Tasks: T15996, T14279

Differential Revision: https://phabricator.grailbio.com/D25229

fbshipit-source-id: 78040e6