lims/ui: convert full page tables to full bleed
Summary:
A number of paged tables are limited in height to 400 or 600 px
depending on width. In addition, they are wrapped in a card that wastes
a bit of space.
Convert pages to full bleed (edge-to-edge), and use most of the vertical
space for table content. Ensure that the table content scrolls without
the header or footer.
Adjust some styling for consistency.
Fix T11621

Test Plan:
Opened batches, conditions, containers, sequencing runs and other pages
that are full page table UIs.
Also spot check other pages that use PagedTable, or use margin from
`page-container` to make sure they work.

{F787674}
{F787673}
{F787672}
{F789366}
{F789365}

Reviewers: O14 Skynet UI, O13 UI Common, O12 LIMS UI, margi, nsawas, lkong, ecarrel

Reviewed By: O14 Skynet UI, O13 UI Common, O12 LIMS UI, nsawas, lkong

Subscribers: mbates, cji, margi, lkong

Maniphest Tasks: T11621

Differential Revision: https://phabricator.grailbio.com/D27929

fbshipit-source-id: 7ef3769