common-components: links are clickable inside two column grids
Summary:
Previously, Two Column Grids used padding to separate rows from each
other. Padding notably does not interact with other elements' padding,
so the padding was set to an excessively high number so that each
row's padding would push all the way to each adjacent row's actual
border, and even a bit past it due to negative margins. This resulted
in this padding being overlayed on top of other rows, making anything
that was supposed to be clickable inside of a row nearly-unclickable.

This changes all the padding to margins. Margins do interact with other
elements' margins, so we only need about half as much to accomplish
the same thing, and each row's margin does not interfere with the
content of other rows. It also gets rid of the negative margins that
contributed to all this behavior.

No real visual changes (from my limited testing)-- but its very
possible this could affect complicated TwoColumnGrids in ways I am
unaware of-- any ideas of places in the UI I should look at?

Fix T19904

Test Plan:
- Spot checked some two column grids and verified they appeared
  approximately identical to before (see images).
- Verified that links in two column grids are now easily clickable.
- Ideas for where in the codebase I should check? Are there complicated
  TwoColumnGrids anywhere I should look at?
{F811565}

{F811564}

{F811563}

{F811562}

{F811561}

{F811560}

Reviewers: nsawas, margi, jzhao, jsingh, O13 UI Common

Reviewed By: jsingh, O13 UI Common

Subscribers: cji

Maniphest Tasks: T19904

Differential Revision: https://phabricator.grailbio.com/D29218

fbshipit-source-id: ba905e4