lims/ui: add part number selector
Summary:
Adds a `PartNumberCommonSelect` component and updates the `InputList`
component so that it can support a generic component rather than just
a `TextInput`, and then uses the `PartNumberCommonSelect` component
inside the `InputList` in the part number details page as a proof-of-
concept.

@jsingh: you're totally correct that on this page, it would make more
sense as just a multi-select since the `InputList` only has one column.
However, I will need to use this
`PartNumberCommonSelect` inside of an `InputList` as part of a new
form I am building that will have multiple columns each of which apply
to a single part number. The part number details page that this diff
updates doesn't actually matter too much
since it's going away soon so I am using it as a proof-of-concept here
to reduce the size of my next diff which I will stack on this one.

Ref T21692

Test Plan:
Clicked around the UI a bunch. Here's pics:
{F845815}
{F845805}

Reviewers: nsawas, margi, jsingh, jzhao, O13 UI Common, O12 LIMS UI

Reviewed By: jsingh, O13 UI Common, O12 LIMS UI

Subscribers: cji, jsingh

Maniphest Tasks: T21692

Differential Revision: https://phabricator.grailbio.com/D31141

fbshipit-source-id: 26c22ab