ui/components: support scrolling for tabbed page
Summary:
To support an overflow of tabs in TabbedLimsPageV2, this revision
defines the Tabs component to be scrollable, and adds an inset
shadow to any side(s) of the tab bar that has scroll buttons.

This revision also fixes and reorganizes some styling.

Fix T22072

Test Plan:
Implement on a test page with multiple tabs

{F852349}

Reviewers: O13 UI Common, jsingh, margi, ecarrel

Reviewed By: O13 UI Common, ecarrel

Subscribers: jzhao, jsingh, cji

Maniphest Tasks: T22072

Differential Revision: https://phabricator.grailbio.com/D31398

fbshipit-source-id: d70b823