lims/ui: paged table export button shows modal with options
Summary:
Replaces the "Export to CSV" button with an "Export" button which pops
up a modal with a list of options. In addition to the user being able
to choose a separator (e.g. tab for a TSV, comma for a CSV, or a
custom value), the user can also choose to download only selected rows,
only visible rows (e.g. all rows on that page regardless of whether
they are checked or not), or all rows in the database satisfying the
given query filter. If the user selects the last option, the UI will
make an additional request to the database to get all the rows (up to
a max of 100,000 for safety).

This also expands the PagedTable types to make it easier for devs to
define custom `exportAccessors` to specify what value will get printed
in the export file for that column.

Ref T21283
Ref T22453
Fix T22643

Test Plan:
- Tested manually in the UI:
{F858633}
- `paged-table.spec.js`

Reviewers: jsingh, nsawas, rhuang, jzhao, O13 UI Common, O12 LIMS UI

Reviewed By: nsawas, rhuang, O13 UI Common, O12 LIMS UI

Maniphest Tasks: T21283, T22453, T22643

Differential Revision: https://phabricator.grailbio.com/D31748

fbshipit-source-id: c659e07