ui/common-components: add wide option to simple table
Summary:
This adds an `isWide` prop to `LargeSimpleTable`. Although the
internals are different, tables with this prop enabled effectively
appear identical to tables without it enabled except that it
expects columns to have defined widths (although they defaults
to 150px) and allows left-right scrolling rather than squeezing
in all columns into the viewable area, which becomes
unreasonable for e.g. sample data keys which can have
any unbounded number of columns. Any number of the leftmost
columns can be defined as "frozen", meaning that they stay fixed to
the left side when the user scrolls to the right (just like in spreadsheets).

Also uses this prop in LIMS in the sample data key table and
purposefully gives bool input fields a little bit of extra space
because they need it.

This is necessary for 2.1.0 because I added a bool input field in the
sample data key table for a plasma isolation step that is 80% obscured
and therefore unusable unless the window is ~4000px wide. This data
key will be set by operators through LIMS, not through Skynet, and so
users must be able to click it.

Ref T22807

Test Plan:
- `large-simple-table.spec.js`
- LIMS UI - Before:
{F871405}
- LIMS UI - After:
{F871406}
- Styleguide (for some reason scrolling is delayed in styleguide but it's instant in LIMS UI):
{F871401}

Reviewers: jsingh, nsawas, rhuang, O13 UI Common, O12 LIMS UI, jzhao

Reviewed By: jsingh, nsawas, O13 UI Common, O12 LIMS UI

Subscribers: cji

Maniphest Tasks: T22807

Differential Revision: https://phabricator.grailbio.com/D32462

fbshipit-source-id: 3999b40