lims/ui: add batches timeline
Summary:
Adds a timeline to the batches view based on the operations API I added
in D33215.

In the future I plan to make the timeline wider and redesign each
element so that the old/new values appear side by side instead of
stacked on top of each other, so that we can put all the values in a
table and thus so they can fit in a smaller vertical area while
expanding into a wider horizontal style. That will be a bigger change
since it will require a new prop `horizontalMode` on the common
component to enable it, so I will add that in a separate diff.

Depends on D33215
Ref T23381

Test Plan: - Looked at it in the UI: {F885473}

Reviewers: rhuang, nsawas, margi, O12 LIMS UI, jsingh, O13 UI Common

Reviewed By: rhuang, O12 LIMS UI, jsingh, O13 UI Common

Subscribers: jsingh, cji

Maniphest Tasks: T23381

Differential Revision: https://phabricator.grailbio.com/D33220

fbshipit-source-id: 63577b9