lims/ui: show deprecated option for assay select
Summary:
There are times when a LIMS user may want to see deprecated assays,
but currently our assay selector hides assays from the common
select when their isActive field is set to false. This revision adds
an item to the bottom of the common select that enables a user to
show deprecated assays.

If searching by text:
{F895835}

If searching by dropdown:
{F895836}

Test Plan:
Go to the batches page, open the omni and click in the Assay selector,
and type "Merlin". No results should be found, but there should be an
option to show deprecated assays. Click it and "cfDNA R&D (Merlin WGS)"
should appear and be selectable.

Reviewers: jsingh, alice, cji, O13 UI Common, O12 LIMS UI, nsawas

Reviewed By: jsingh, O13 UI Common, O12 LIMS UI

Maniphest Tasks: T15931

Differential Revision: https://phabricator.grailbio.com/D33735

fbshipit-source-id: 7596492