common/components: add noOptionsOverride to multi-select
Summary:
This revision adds an extension to the CommonMultiSelect component to
allow developers to override the message that displays when no options
are found matching the entered text (the default is "No results found",
which is sometimes misleading).

For example, T17745 raised the issue that when adding sample labels
to a new issue form input, if a duplicate sample is created, "No results
found" is misleading. It would be more appropriate to say "This label
has already been entered".

Test Plan: Should not break the `common-multi-select.spec.js` cypress test.

Reviewers: O13 UI Common, O12 LIMS UI, nsawas, jsingh, ecarrel

Reviewed By: O13 UI Common, O12 LIMS UI, nsawas, ecarrel

Subscribers: jsingh, cji

Maniphest Tasks: T17745

Differential Revision: https://phabricator.grailbio.com/D34308

fbshipit-source-id: 344e179