ui/components: fix flex styling for PagedTable
Summary:
Noticed in Safari (not seen in Chrome) the PagedTable has styling issues
because the table content size is ambiguous. Defining `flex: 1` allows
the browser to know that the size of the table content should fit the rest
of the screen, giving proper room for the header and the footer.

Fix T22317

Test Plan: Test change in chrome inspector

Reviewers: O13 UI Common, jsingh, margi

Reviewed By: O13 UI Common, margi

Subscribers: cji

Maniphest Tasks: T22317

Differential Revision: https://phabricator.grailbio.com/D35232

fbshipit-source-id: c45a43c