lims/ui: filter pipeline run storage locations by flow cell run
Summary:
When kicking off pipeline runs, users can specify the storage
bucket for the run data. It doesn't make sense to kick off a
pipeline run in a different location than where the flowcell
run data that is triggering the run is stored. Therefore, this
revision automatically filters the storage location dropdown
options to only include those that are associated with the flowcell
run ids listed in the "Specification Details" input on the same page.

First diff on T25510.

Test Plan:
Tested and verified in UI:

{F927704}

Reviewers: rhuang, jsingh, O13 UI Common, O12 LIMS UI, nsawas

Reviewed By: jsingh, O13 UI Common, O12 LIMS UI

Subscribers: ecarrel

Maniphest Tasks: T25510

Differential Revision: https://phabricator.grailbio.com/D35527

fbshipit-source-id: d772d29