ui/components: fix call limit exceeded
Summary:
There is a know bug with the observer that does not cause any performance issues
however produces a lot of errors in the console. It is properly tracked
here:

https://github.com/WICG/ResizeObserver/issues/38

By disconnected the observer between calls and requesting the specific animation
frame to be ovserved we can prevent the error from happening.

Ref T15764

Test Plan: Test resizing UI manually

Reviewers: O13 UI Common, jsingh, margi

Reviewed By: O13 UI Common, margi

Subscribers: cji

Maniphest Tasks: T15764

Differential Revision: https://phabricator.grailbio.com/D37310

fbshipit-source-id: 01cc950