ui/common-components: support isVisible prop to fix crash on reagent lot page
Summary:
The `isVisible` prop is defined as part of the `PageConfigV2` flow type
but was never supported. Unfortunately, I did not realize this when
using this on the reagent lot detail page, so this means that tabs
might show up for lots even if that tab does not apply to those lots
and is not supported, causing the UI to crash.

Consequently, this should be cherrypicked into 2.9.0.

Ref T24942
Ref T15764

Test Plan: - Tested reagent lot detail page in the UI

Reviewers: nsawas, arinehouse, rhuang, O13 UI Common

Reviewed By: rhuang, O13 UI Common

Subscribers: cji

Maniphest Tasks: T24942, T15764

Differential Revision: https://phabricator.grailbio.com/D37910

fbshipit-source-id: 1f25699