ui/components: add autosizeHeight prop to LargeSimpleTable
Summary:
Adds a new prop, `autosizeHeight`, to the `LargeSimpleTable` common
component. This allows the caller to autosize the height of the table
to be 100% the height of its container element.

Needed for T28297.

Test Plan:
Tested and verified in UI. Works with existing version of the table as
the prop defaults to false, which maintains the same behavior. The
functionality will be demonstrated in another diff (on T28297).

Reviewers: ecarrel, lkong, nsawas, O13 UI Common

Reviewed By: ecarrel, O13 UI Common

Subscribers: cji

Maniphest Tasks: T28297

Differential Revision: https://phabricator.grailbio.com/D38050

fbshipit-source-id: e6fe491