common/components: fix height on LargeWideTableComponent
Summary:
The LargeWideTableComponent was updated in a previous revision to
allow it to stretch and fill an entire container div. I added a field
that prevented overflow in order to accomplish this. However, this
made the height of the table inaccurate and hid the bottom row.

This revision fixes the component.

Note that because this is just a patch for a previous revision, I did
not update the changelog.

Test Plan: Tested in UI.

Reviewers: ecarrel, jzhao, nsawas, O13 UI Common

Reviewed By: jzhao, O13 UI Common

Subscribers: cji

Maniphest Tasks: T28297

Differential Revision: https://phabricator.grailbio.com/D38145

fbshipit-source-id: 19652cd