ui: use eslint-prettier plugin for formatting code
Summary:
We currently use eslint and prettier seperately and have formatting rules in each.
There are conflicting rules and the process tend to be slow.  We use the recommended integration to combine them.
Also change prettier line-width to 80 because it is a better default (even with larger screens we may not want longer lines).

Test Plan: UI

Reviewers: ecarrel, arinehouse, nsawas, lkong, O13 UI Common, O12 LIMS UI

Reviewed By: nsawas, O13 UI Common, O12 LIMS UI

Maniphest Tasks: T8000

Differential Revision: https://phabricator.grailbio.com/D38377

fbshipit-source-id: affdc98