common-components: fix types and disabled appearance of dropdown menus
Summary:
I tried to land these as part of D39640, but somehow right before
landing, I screwed up git and all my changes except the ones in
`dist` got thrown out. This is just the same as the changes that were
already approved in D39640.

- Disabled menu items should look disabled
- Fixes incorrect type which incorrectly suggested that `href` was a
  supported attribute when currently it is not

Ref T15764

Test Plan:
Looked in the UI at a lot and tried to set its retains without having
permissions to do so, saw that it appeared greyed out

Reviewers: O13 UI Common, nsawas, arinehouse, jzhao

Reviewed By: O13 UI Common, nsawas

Subscribers: cji

Maniphest Tasks: T15764

Differential Revision: https://phabricator.grailbio.com/D39643

fbshipit-source-id: 2ec7424