common/components: fix z-indexing for sign in page
Summary:
Fixes an issue where the Autocomplete dropdown on the sign in page
was not displaying correctly. This was due to the z-index of the
SignInPage component being too high (9999) which seems to have been
totally unnecessary, as no components should be rendering behind it.

Test Plan: Tested in UI

Reviewers: jzhao, O13 UI Common, nsawas

Reviewed By: jzhao, O13 UI Common

Subscribers: cji

Tags: #platform-wide

Maniphest Tasks: T30742

Differential Revision: https://phabricator.grailbio.com/D40478

fbshipit-source-id: 11a58cc