lims/ui: Allow comma seperated list for common suggest
Summary: CommonSuggest use to be able to handle a comma seperated list of options.  While this is still the case, the options menu won't show up after the first item.  Here we revert to the old common suggest behavior where the value is parsed as a CSV and the last item is used as input match.

Test Plan: {F1040701}

Reviewers: ecarrel, nsawas, arinehouse, O13 UI Common, O12 LIMS UI

Reviewed By: nsawas, O13 UI Common, O12 LIMS UI

Tags: #software-only_change_control, #platform-wide

Maniphest Tasks: T31735

Differential Revision: https://phabricator.grailbio.com/D41421

fbshipit-source-id: ba7b1fc